♻️ refactor: header button handleClick 수정
